chore: remove project billing
