+++ packages/ee/shared/src/index.ts
@@ -1,5 +1,4 @@
 export * from './lib/billing';
-export * from './lib/billing/project-billing';
 export * from './lib/audit-events';
 export * from './lib/git-repo';
 export * from './lib/api-key';
