--- packages/react-ui/src/app/routes/plans/index.tsx
@@ -1,5 +1,0 @@
-import { Plans } from '@/features/plans/components/plans';
-
-export default function PlansPage() {
-  return <Plans />;
-}
