+++ packages/server/api/src/app/ee/projects/platform-project-service.ts
@@ -33,11 +33,11 @@ import { system } from '../../helper/system/system'
 import { ProjectEntity } from '../../project/project-entity'
 import { projectService } from '../../project/project-service'
 import { userService } from '../../user/user-service'
-import { projectBillingService } from '../billing/project-billing/project-billing.service'
 import { BillingEntityType, usageService } from '../platform-billing/usage/usage-service'
 import { ProjectMemberEntity } from '../project-members/project-member.entity'
 import { projectLimitsService } from '../project-plan/project-plan.service'
 import { platformProjectSideEffects } from './platform-project-side-effects'
+import { platformBillingService } from '../platform-billing/platform-billing.service'
 const projectRepo = repoFactory(ProjectEntity)
 const projectMemberRepo = repoFactory(ProjectMemberEntity)
 
@@ -138,7 +138,7 @@ async function getProjects(params: GetAllParams & { projectIds?: string[] }, log
         ...spreadIfDefined('displayName', displayName),
         ...(projectIds ? { id: In(projectIds) } : {}),
     }
-   
+
     const queryBuilder = projectRepo()
         .createQueryBuilder('project')
         .leftJoinAndMapOne(
@@ -148,8 +148,8 @@ async function getProjects(params: GetAllParams & { projectIds?: string[] }, log
             'project.id = "project_plan"."projectId"',
         )
         .where(filters)
-        .groupBy('project.id') 
-        .addGroupBy('"project_plan"."id"') 
+        .groupBy('project.id')
+        .addGroupBy('"project_plan"."id"')
 
     const { data, cursor } = await paginator.paginate(queryBuilder)
     const projects: ProjectWithLimits[] = await Promise.all(
@@ -179,8 +179,9 @@ async function isSubscribedInStripe(projectId: ProjectId, log: FastifyBaseLogger
     if (!isCloud) {
         return false
     }
-    const status = await projectBillingService(log).getOrCreateForProject(projectId)
-    return status.subscriptionStatus === ApSubscriptionStatus.ACTIVE
+    const project = await projectService.getOneOrThrow(projectId)
+    const status = await platformBillingService(log).getOrCreateForPlatform(project.platformId)
+    return status.stripeSubscriptionStatus === ApSubscriptionStatus.ACTIVE
 }
 function isCustomerPlatform(platformId: string | undefined): boolean {
     if (isNil(platformId)) {
@@ -202,7 +203,7 @@ async function enrichProject(
         .groupBy('user.id')
         .where(`user.status = '${UserStatus.ACTIVE}' and project_member."projectId" = '${project.id}'`)
         .getCount()
-  
+
     const totalFlows = await flowService(log).count({
         projectId: project.id,
     })
@@ -212,7 +213,7 @@ async function enrichProject(
         status: FlowStatus.ENABLED,
     })
 
-  
+
     return {
         ...project,
         plan: await projectLimitsService.getOrCreateDefaultPlan(
@@ -223,7 +224,7 @@ async function enrichProject(
             project.id,
             BillingEntityType.PROJECT,
         ),
-        analytics: { 
+        analytics: {
             activeFlows,
             totalFlows,
             totalUsers,
