+++ packages/server/api/src/assets/emails/quota-100.html
@@ -498,7 +498,7 @@
                                                     "
                                                   >
                                                     <a
-                                                      href="https://cloud.activepieces.com/plans"
+                                                      href="https://cloud.activepieces.com/platform/setup/billing"
                                                       target="_blank"
                                                       class="link-button button-inner"
                                                       style="
